Moved some components out, added icon images
